requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
pymysql>=1.1.0
python-docx==1.1.0
docxtpl==0.16.7
APScheduler==3.10.4
python-dateutil==2.8.2
docx2pdf==0.1.8
requests==2.31.0
python-dotenv==1.0.0

